Name the t2v test files after the model each one runs
