+++ integrations_v2/t2v_wan21/t2v_wan21/tests/test_stand_in_model.py
@@ -1,20 +1,23 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""CPU tests for the Cosmos Predict2 application, against a stand-in model.
+"""CPU tests for the Wan 2.1 application, against a stand-in model.
 
 What is specific to this integration is which model it runs and that the model
 generates its clip in one block, so that is what these cover. How a
 text-to-video application behaves in general belongs to the shared layer and is
 covered in ``flashdreams/test_v2``, which is why there is so little here.
+
+The one thing a stand-in cannot show is what the checkpoint generates, which is
+what ``test_real_model.py`` alongside this is for. Nothing here needs a GPU.
 """
 
 import shutil
 from pathlib import Path
 
 import pytest
-from cosmos_predict2.config import RUNNER_COSMOS2_T2V_2B_720P
-from t2v_cosmos_predict2 import CosmosPredict2T2VApplication
+from t2v_wan21 import Wan21T2VApplication
+from wan21.config import RUNNER_WAN21_T2V_1PT3B_480P
 
 from flashdreams.runtime_v2.session_desc import SessionDesc
 from flashdreams.runtime_v2.video_tensor import VideoTensorLayout
@@ -36,20 +39,22 @@ def test_the_model_says_what_it_generates_without_being_told() -> None:
 
     They come from the runner config this integration ships, which is the point
     of deriving the defaults from it: a caller wanting the clip the model was
-    trained to generate passes no size flags at all.
+    trained to generate passes no size flags at all. The rollout length is the
+    exception, since a config for a model that does not roll out does not carry
+    one.
     """
-    app = CosmosPredict2T2VApplication(pipeline_config=FakeT2VPipelineConfig())
+    app = Wan21T2VApplication(pipeline_config=FakeT2VPipelineConfig())
     app.init(["--prompt", _PROMPT, "--device", "cpu"])
 
     desc = app.session_desc()
 
     assert (desc.video_width, desc.video_height) == (
-        RUNNER_COSMOS2_T2V_2B_720P.pixel_width,
-        RUNNER_COSMOS2_T2V_2B_720P.pixel_height,
+        RUNNER_WAN21_T2V_1PT3B_480P.pixel_width,
+        RUNNER_WAN21_T2V_1PT3B_480P.pixel_height,
     )
-    assert desc.frames_per_second_for_step == RUNNER_COSMOS2_T2V_2B_720P.fps
+    assert desc.frames_per_second_for_step == RUNNER_WAN21_T2V_1PT3B_480P.fps
     assert desc.output_layout is VideoTensorLayout.tchw
-    assert app.total_blocks == RUNNER_COSMOS2_T2V_2B_720P.total_blocks == 1
+    assert app.total_blocks == 1
 
 
 @pytest.mark.parametrize("total_blocks", [2, 60])
@@ -57,7 +62,7 @@ def test_a_rollout_is_refused_because_this_model_does_not_roll_out(
     total_blocks: int,
 ) -> None:
     """A second block would not continue the first, so asking is a mistake."""
-    app = CosmosPredict2T2VApplication(pipeline_config=FakeT2VPipelineConfig())
+    app = Wan21T2VApplication(pipeline_config=FakeT2VPipelineConfig())
 
     with pytest.raises(ValueError, match="must be 1"):
         app.init(
@@ -81,12 +86,12 @@ def test_a_run_writes_every_generated_frame_to_an_mp4(tmp_path: Path) -> None:
     path = tmp_path / "clip.mp4"
 
     result = check_t2v_model_impl(
-        CosmosPredict2T2VApplication(pipeline_config=FakeT2VPipelineConfig(pipeline)),
+        Wan21T2VApplication(pipeline_config=FakeT2VPipelineConfig(pipeline)),
         # The stand-in generates its own size rather than the checkpoint's, so
         # it says so here rather than asking the application.
         SessionDesc(
             output_layout=VideoTensorLayout.tchw,
-            frames_per_second_for_step=RUNNER_COSMOS2_T2V_2B_720P.fps,
+            frames_per_second_for_step=RUNNER_WAN21_T2V_1PT3B_480P.fps,
             video_width=pipeline.width,
             video_height=pipeline.height,
         ),
